FR-45: Reestructure folders
