+++ src/Components/Members/Table/table.modules.css
@@ -1,0 +1,43 @@
+table {
+  border-collapse: collapse;
+}
+
+table th, table td {
+  padding: 15px 12px;
+  margin: 0 10px;
+  width: 5%;
+  box-sizing: border-box;
+  text-align: left;
+}
+
+.row {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+
+button {
+  text-align: center;
+  vertical-align: baseline;
+  margin: 20px 25px 5px 25px;
+  padding: 5px;
+}
+
+.close_btn {
+  text-align: center;
+  color: #fff;
+  background-color: red;
+}
+
+.close_btn:hover,
+.close_btn:focus {
+  color: #000;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.edit_btn {
+  text-align: center;
+  color: #fff;
+  background-color: red
+}
